docshell/test/mochitest/test_bug598895.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=598895
-->
<head>
  <title>Test for Bug 598895</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/WindowSnapshot.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=598895">Mozilla Bug 598895</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 598895 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
var win1 = window.open();
win1.document.body.textContent = "Should show";

var windowsLoaded = 0;

window.onmessage = function(ev) {
  is(ev.data, "loaded", "Message should be 'loaded'");
  if (++windowsLoaded == 2) {
    var one = snapshotWindow(win1);
    var two = snapshotWindow(win2);
    var three = snapshotWindow(win3);
    win1.close();
    win2.close();
    win3.close();
    ok(compareSnapshots(one, two, true)[0], "Popups should look identical");
    ok(compareSnapshots(one, three, false)[0], "Popups should not look identical");

    SimpleTest.finish();
  }
};

var win2 = window.open("file_bug598895_1.html");
var win3 = window.open("file_bug598895_2.html");
});
</script>
</pre>
</body>
</html>
